+++ gherkin/javascript/src/pickles/compile.ts
@@ -79,12 +79,18 @@ function compileRule(
   rule.children.forEach((stepsContainer) => {
     if (stepsContainer.background) {
       ruleBackgroundSteps = ruleBackgroundSteps.concat(stepsContainer.background.steps)
+    } else if (stepsContainer.scenario.examples.length === 0) {
+      compileScenario(
+        tags,
+        ruleBackgroundSteps,
+        stepsContainer.scenario,
+        language,
+        pickles,
+        uri,
+        newId
+      )
     } else {
-      const compileFunc = stepsContainer.scenario.examples.length === 0 
-        ? compileScenario
-        : compileScenarioOutline
-      
-      compileFunc(
+      compileScenarioOutline(
         tags,
         ruleBackgroundSteps,
         stepsContainer.scenario,